[af-263] refactor: 매장 재고 환불 package 위치 변경
