+++ src/main/java/com/drunkenlion/alcoholfriday/domain/admin/restaurant/refund/dao/RestaurantOrderRefundDetailRepository.java
@@ -1,6 +1,6 @@
-package com.drunkenlion.alcoholfriday.domain.restaurant.refund.dao;
+package com.drunkenlion.alcoholfriday.domain.admin.restaurant.refund.dao;
 
-import com.drunkenlion.alcoholfriday.domain.restaurant.refund.entity.RestaurantOrderRefundDetail;
+import com.drunkenlion.alcoholfriday.domain.admin.restaurant.refund.entity.RestaurantOrderRefundDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RestaurantOrderRefundDetailRepository extends JpaRepository<RestaurantOrderRefundDetail, Long> {
