+++ src/main/java/com/drunkenlion/alcoholfriday/domain/admin/restaurant/refund/util/RestaurantOrderRefundStatusConverter.java
@@ -1,6 +1,6 @@
-package com.drunkenlion.alcoholfriday.domain.restaurant.refund.util;
+package com.drunkenlion.alcoholfriday.domain.admin.restaurant.refund.util;
 
-import com.drunkenlion.alcoholfriday.domain.restaurant.refund.enumerated.RestaurantOrderRefundStatus;
+import com.drunkenlion.alcoholfriday.domain.admin.restaurant.refund.enumerated.RestaurantOrderRefundStatus;
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
